+++ packages/type/src/Rules/__tests__/BigIntValidationRule.spec.ts
@@ -1,4 +1,4 @@
-import { BigIntValidationRule } from '../BigIntValidationRule';
+import { BigIntValidationRule } from '../BigIntValidationRule.js';
 
 describe('BigIntValidationRule', () => {
   describe('evaluate', () => {